+++ crates/matrix-sdk/src/client/caches.rs
@@ -12,17 +12,44 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use matrix_sdk_base::ttl_cache::TtlCache;
-use ruma::api::client::discovery::get_authorization_server_metadata::v1::AuthorizationServerMetadata;
+use matrix_sdk_base::{store::WellKnownResponse, ttl_cache::TtlCache};
+use ruma::api::{
+    SupportedVersions,
+    client::discovery::get_authorization_server_metadata::v1::AuthorizationServerMetadata,
+};
 use tokio::sync::RwLock;
-
-use super::ClientServerInfo;
 
 /// A collection of in-memory data that the `Client` might want to cache to
 /// avoid hitting the homeserver every time users request the data.
 pub(crate) struct ClientCaches {
-    /// Server info, either prefilled during building or fetched from the
+    /// Supported versions, either prefilled during building or fetched from the
     /// server.
-    pub(super) server_info: RwLock<ClientServerInfo>,
+    pub(super) supported_versions: RwLock<CachedValue<SupportedVersions>>,
+    /// Well-known information.
+    pub(super) well_known: RwLock<CachedValue<Option<WellKnownResponse>>>,
     pub(crate) server_metadata: tokio::sync::Mutex<TtlCache<String, AuthorizationServerMetadata>>,
+}
+
+/// A cached value that can either be set or not set, used to avoid confusion
+/// between a value that is set to `None` (because it doesn't exist) and a value
+/// that has not been cached yet.
+#[derive(Clone)]
+pub(super) enum CachedValue<Value> {
+    /// A value has been cached.
+    Cached(Value),
+    /// Nothing has been cached yet.
+    NotSet,
+}
+
+impl<Value> CachedValue<Value> {
+    /// Takes the value out of the `CachedValue`, leaving a `NotSet` in its
+    /// place.
+    pub(super) fn take(&mut self) -> Option<Value> {
+        let prev = std::mem::replace(self, Self::NotSet);
+
+        match prev {
+            Self::Cached(value) => Some(value),
+            Self::NotSet => None,
+        }
+    }
 }